Fix compiler warnings
* Disable MinGW warnings in thirdparty/opus reg. unused parameters
* Fix MSVC compiler warnings, reg. signed/unsigned mismatch (C4245)
* Enable some of the previously (and apparently unnecessarily) disabled MSVC compiler warnings in src/engraving again
* Fix MSVC compiler warnings reg. conversion from `size_t` to `int`, outside of engraving

+++ src/framework/audio/internal/synthesizers/fluidsynth/sfcachedloader.h
@@ -92,7 +92,7 @@ void* openSoundFont(const char* filename)
 
 int readSoundFont(void* buf, int count, void* handle)
 {
-    return std::fread(buf, count, 1, static_cast<std::FILE*>(handle));
+    return static_cast<int>(std::fread(buf, count, 1, static_cast<std::FILE*>(handle)));
 }
 
 int seekSoundFont(void* handle, long offset, int origin)
